Refactor popup window content to a common base CommandMenu has popup selector and fallback to original in double click
