+++ Assets/Fungus/Scripts/Editor/BasePopupWindowContent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be9eead8f19216444b4a87e94fe02ed3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
